j'ajouté la partie qui permet l'autheur d'ajouter les tag a chaque wiki

+++ app/Controllers/CategoriesController.php
@@ -12,54 +12,84 @@ class CategoriesController
 
     public static function categories()
     {
-        $categorie = new CategorieModel();
-        $categories = $categorie->fetchAllCategories();
+        if (controller::isAdmin()) {
+            $categorie = new CategorieModel();
+            $categories = $categorie->fetchAllCategories();
 
-        $nbrCategories = $categorie->countCategories();
-        require __DIR__ . "/../../View/Dashboard/Categories.php";
+            $nbrCategories = $categorie->countCategories();
+            require __DIR__ . "/../../View/Dashboard/Categories.php";
+        } else {
+            header('location: /home');
+
+        }
+
     }
 
     public static function Addcategories()
     {
+        if (controller::isAdmin()) {
+
             require __DIR__ . "/../../View/Dashboard/Addcategories.php";
+        } else {
+            header('location: /home');
+
+        }
     }
 
     public static function saveCategorie()
     {
-        if(isset($_POST))
-        {
-            $addcategorie = new CategorieModel();
-            $addcategorie->addCategories($_POST);
-            header('location: /../../Addcategories');
+        if (controller::isAdmin()) {
+
+            if (isset($_POST)) {
+                $addcategorie = new CategorieModel();
+                $addcategorie->addCategories($_POST);
+                header('location: /../../Addcategories');
+            }
+        } else {
+            header('location: /home');
+
         }
     }
 
     public static function deleteCategorie()
     {
-        $id = $_GET['id'];
-        $deletecategorie = new CategorieModel();
-        $deletecategorie->deleteCatego($id);
-        header("location: /../../categories");
+        if (controller::isAdmin()) {
+
+            $id = $_GET['id'];
+            $deletecategorie = new CategorieModel();
+            $deletecategorie->deleteCatego($id);
+            header("location: /../../categories");
+        } else {
+            header('location: /home');
+
+        }
     }
 
 
     public static function formUpdateCategorie()
     {
-        $id = $_GET['id'];
-        $categorieObj = new CategorieModel();
-        $categorie = $categorieObj->getCatgorie($id);
-        require __DIR__ . "/../../View/Dashboard/updateCategories.php";
+        if (controller::isAdmin()) {
+            $id = $_GET['id'];
+            $categorieObj = new CategorieModel();
+            $categorie = $categorieObj->getCatgorie($id);
+            require __DIR__ . "/../../View/Dashboard/updateCategories.php";
+        } else {
+            header('location: /home');
 
+        }
     }
 
     public static function submitupdateCategorie()
     {
-        $objCatgorie = new CategorieModel();
-        $objCatgorie->submitUpdateCatgorie($_POST);
-        header("location: /../../categories");
+        if (controller::isAdmin()) {
+            $objCatgorie = new CategorieModel();
+            $objCatgorie->submitUpdateCatgorie($_POST);
+            header("location: /../../categories");
+        } else {
+            header('location: /home');
 
+        }
     }
-
 
 
 }
