+++ app/Controllers/TagsController.php
@@ -10,49 +10,72 @@ class TagsController
 
     public static function tags()
     {
-        $tag = new TagsModel();
-        $tags = $tag->fetchAllCategories();
+        if(controller::isAdmin()){
+            $tag = new TagsModel();
+            $tags = $tag->fetchAllTags();
 
-        $nbrtags = $tag->countTags();
-        require __DIR__ . "/../../View/Dashboard/Tags.php";
+            $nbrtags = $tag->countTags();
+            require __DIR__ . "/../../View/Dashboard/Tags.php";
+        }else{
+            header('location: /home');
+        }
     }
 
     public static function addTag()
     {
-        require __DIR__ . "/../../View/Dashboard/AddTags.php";
+        if(controller::isAdmin()){
+            require __DIR__ . "/../../View/Dashboard/AddTags.php";
+        }else{
+            header('location: /home');
+        }
     }
 
     public static function deletetag()
     {
-        $id = $_GET['id'];
-        $deletetag = new TagsModel();
-        $deletetag->deleteTag($id);
-        header("location: /../../tags");
+        if(controller::isAdmin()){
+            $id = $_GET['id'];
+            $deletetag = new TagsModel();
+            $deletetag->deleteTag($id);
+            header("location: /../../tags");
+        }else{
+            header('location: /home');
+        }
     }
 
     public static function  saveTags()
     {
-        if(isset($_POST))
-        {
-            $addtags = new TagsModel();
-            $addtags->addTags($_POST);
-            header('location: /../../Addtags');
+        if(controller::isAdmin()){
+            if(isset($_POST))
+            {
+                $addtags = new TagsModel();
+                $addtags->addTags($_POST);
+                header('location: /../../Addtags');
+            }
+        }else{
+            header('location: /home');
         }
     }
 
     public static function formUpdateTags()
     {
-        $id = $_GET['id'];
-        $tagObj = new TagsModel();
-        $tag = $tagObj->getTag($id);
-        require __DIR__ . "/../../View/Dashboard/updateTags.php";
+        if(controller::isAdmin()){
+            $id = $_GET['id'];
+            $tagObj = new TagsModel();
+            $tag = $tagObj->getTag($id);
+            require __DIR__ . "/../../View/Dashboard/updateTags.php";
+        }else{
+            header('location: /home');
+        }
     }
 
     public static function submitUpdateTags()
     {
-        $objTags = new TagsModel();
-        $objTags->submitTags($_POST);
-        header("location: /../../tags");
+        if(controller::isAdmin()){
+            $objTags = new TagsModel();
+            $objTags->submitTags($_POST);
+            header("location: /../../tags");
+        }else{
+            header('location: /home');
+        }
     }
-
 }
